+++ src/test/converter2/issues/gh2291.ts
@@ -1,0 +1,14 @@
+/**
+ * Parses a JSON string to the specified type.
+ */
+export function safeParse<T>(
+    guard: (o: unknown) => o is T,
+    json: string,
+): T | undefined {
+    try {
+        const o = JSON.parse(json) as unknown;
+        return guard(o) ? o : undefined;
+    } catch {
+        return undefined;
+    }
+}