+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 ### Bug Fixes
 
+-   `notDocumented` validation will no longer require documentation for data within parameters that cannot be documented via `@param`, #2291.
 -   Type parameters will now be resolved for arrow-methods on classes like regular class methods, #2320.
 -   TypeDoc now inherits `typedocOptions` fields from extended tsconfig files, #2334.
 -   Methods which return function types no longer have duplicated comments, #2336.